Refactor/hu 8 auth

+++ src/main/java/com/powerup/user/users/application/dto/response/UserResponse.java
@@ -2,6 +2,6 @@ package com.powerup.user.users.application.dto.response;
 
 import java.time.LocalDate;
 
-public record UserResponse(Long id, String firstName, String lastName, Integer identityDocument, String phoneNumber,
+public record UserResponse(Long id, String firstName, String lastName, Long identityDocument, String phoneNumber,
                            LocalDate birthDate, String email, String password, String role) {
 }
